instead of using :bufdo, use writefile()
currently, we use :bufdo to loop over all buffers and write backup
files.
While this works, it can cause some problems, e.g. :bufdo is not allowed
in the sandbox or the bufdo call might cause the current window to
scroll.

Therefore, explicitly loop over a range of (1,bufnr('$')) and use
writefile() to write the buffer correctly.

This has the drawback however that we cannot cache the latest write, so
we might save the file several times.